refactor(WebChessController.java): service로 수정
컨트롤러 역할이 아닌 서비스 역할을 하고 있는 객체의 이름과 패키지 수정

BREAKING CHANGE: WebChessController의 이름과 패키지 수정

as-is:
/chess/Controller/web/WebChessController.java

to-be:
/chess/service/WebChessService.java
